java: Add FDefaultServerEventHandler
This is ultimately renaming FDefaultNatsServerEventHandler to something
more generic so that it can be used for servlets in a subsequent commit
without causing confusion.
